Added first test of hand menu where 3 cards can be inserted
Would be great if the number of cards the player can have in the hand is configureable. Also it would be great if cards can be inserted in between other cards. Would require detecting the two nearest slots and move the cards between slots to make room. But what should we do if player releases a card and there is no more room?